+++ src/os/bootloader_tools/devicemanager/src/com/ngos/bootloader_tools/devicemanager/other/devicemanagermode.h
@@ -45,7 +45,7 @@ inline const char8* enumToFullString(DeviceManagerMode mode) // TEST: NO
 
     static char8 res[17];
 
-    sprintf(res, "0x%02X (%s)", mode, enumToString(mode));
+    sprintf(res, "0x%02X (%s)", (u8)mode, enumToString(mode));
 
     return res;
 }
