+++ src/os/bootloader_tools/memorytest/src/com/ngos/bootloader_tools/memorytest/other/testmode.h
@@ -45,7 +45,7 @@ inline const char8* enumToFullString(TestMode mode) // TEST: NO
 
     static char8 res[18];
 
-    sprintf(res, "0x%02X (%s)", mode, enumToString(mode));
+    sprintf(res, "0x%02X (%s)", (u8)mode, enumToString(mode));
 
     return res;
 }
