+++ src/os/configure/src/com/ngos/configure/b_early/main/decompressors/xz/lib/lzma2/lzmastate.h
@@ -79,7 +79,7 @@ inline const char8* enumToFullString(LzmaState state) // TEST: NO
 
     static char8 res[35];
 
-    sprintf(res, "0x%02X (%s)", state, enumToString(state));
+    sprintf(res, "0x%02X (%s)", (u8)state, enumToString(state));
 
     return res;
 }
