+++ src/os/kernel/src/com/ngos/kernel/other/ioremap/fixmap.h
@@ -41,7 +41,7 @@ inline const char8* enumToFullString(FixMap map) // TEST: NO
 
     static char8 res[15];
 
-    sprintf(res, "0x%02X (%s)", map, enumToString(map));
+    sprintf(res, "0x%02X (%s)", (u8)map, enumToString(map));
 
     return res;
 }
