+++ src/os/shared/common/src/com/ngos/shared/common/cpu/cpuvendor.h
@@ -47,7 +47,7 @@ inline const char8* enumToFullString(CpuVendor vendor) // TEST: NO
 
     static char8 res[15];
 
-    sprintf(res, "0x%02X (%s)", vendor, enumToString(vendor));
+    sprintf(res, "0x%02X (%s)", (u8)vendor, enumToString(vendor));
 
     return res;
 }
