+++ src/os/shared/common/src/com/ngos/shared/common/cpu/model/intelcpumodel.h
@@ -128,7 +128,7 @@ inline const char8* enumToFullString(CpuFamily family, IntelCpuModel model) // T
 
     static char8 res[41];
 
-    sprintf(res, "0x%02X (%s)", model, enumToString(family, model));
+    sprintf(res, "0x%02X (%s)", (u8)model, enumToString(family, model));
 
     return res;
 }
