+++ src/os/shared/common/src/com/ngos/shared/common/cpu/x86bugword.h
@@ -43,7 +43,7 @@ inline const char8* enumToFullString(x86BugWord word) // TEST: NO
 
     static char8 res[25];
 
-    sprintf(res, "0x%02X (%s)", word, enumToString(word));
+    sprintf(res, "0x%02X (%s)", (u8)word, enumToString(word));
 
     return res;
 }
