+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmibiosextendedromsizeunit.h
@@ -43,7 +43,7 @@ inline const char8* enumToFullString(DmiBiosExtendedRomSizeUnit unit) // TEST: N
 
     static char8 res[17];
 
-    sprintf(res, "0x%02X (%s)", unit, enumToString(unit));
+    sprintf(res, "0x%02X (%s)", (u8)unit, enumToString(unit));
 
     return res;
 }
