+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmibits64memoryerrorinformationerroroperation.h
@@ -51,7 +51,7 @@ inline const char8* enumToFullString(DmiBits64MemoryErrorInformationErrorOperati
 
     static char8 res[21];
 
-    sprintf(res, "0x%02X (%s)", operation, enumToString(operation));
+    sprintf(res, "0x%02X (%s)", (u8)operation, enumToString(operation));
 
     return res;
 }
