+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmibits64memoryerrorinformationerrortype.h
@@ -69,7 +69,7 @@ inline const char8* enumToFullString(DmiBits64MemoryErrorInformationErrorType ty
 
     static char8 res[34];
 
-    sprintf(res, "0x%02X (%s)", type, enumToString(type));
+    sprintf(res, "0x%02X (%s)", (u8)type, enumToString(type));
 
     return res;
 }
