+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmichassiscontainedelementtypeselect.h
@@ -43,7 +43,7 @@ inline const char8* enumToFullString(DmiChassisContainedElementTypeSelect select
 
     static char8 res[22];
 
-    sprintf(res, "0x%02X (%s)", select, enumToString(select));
+    sprintf(res, "0x%02X (%s)", (u8)select, enumToString(select));
 
     return res;
 }
