+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmichassisstate.h
@@ -53,7 +53,7 @@ inline const char8* enumToFullString(DmiChassisState state) // TEST: NO
 
     static char8 res[23];
 
-    sprintf(res, "0x%02X (%s)", state, enumToString(state));
+    sprintf(res, "0x%02X (%s)", (u8)state, enumToString(state));
 
     return res;
 }
