+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmimemorydevicetechnology.h
@@ -55,7 +55,7 @@ inline const char8* enumToFullString(DmiMemoryDeviceTechnology technology) // TE
 
     static char8 res[31];
 
-    sprintf(res, "0x%02X (%s)", technology, enumToString(technology));
+    sprintf(res, "0x%02X (%s)", (u8)technology, enumToString(technology));
 
     return res;
 }
