+++ src/os/shared/common/src/com/ngos/shared/common/dmi/entry/lib/dmivoltageprobestatus.h
@@ -53,7 +53,7 @@ inline const char8* enumToFullString(DmiVoltageProbeStatus status) // TEST: NO
 
     static char8 res[23];
 
-    sprintf(res, "0x%02X (%s)", status, enumToString(status));
+    sprintf(res, "0x%02X (%s)", (u8)status, enumToString(status));
 
     return res;
 }
