+++ src/os/shared/common/src/com/ngos/shared/common/elf/programtype.h
@@ -65,7 +65,7 @@ inline const char8* enumToFullString(ElfProgramType type) // TEST: NO
 
     static char8 res[21];
 
-    sprintf(res, "0x%08X (%s)", type, enumToString(type));
+    sprintf(res, "0x%08X (%s)", (u32)type, enumToString(type));
 
     return res;
 }
