+++ src/os/shared/common/src/com/ngos/shared/common/elf/sectiontype.h
@@ -84,7 +84,7 @@ inline const char8* enumToFullString(ElfSectionType type) // TEST: NO
 
     static char8 res[27];
 
-    sprintf(res, "0x%08X (%s)", type, enumToString(type));
+    sprintf(res, "0x%08X (%s)", (u32)type, enumToString(type));
 
     return res;
 }
