+++ src/os/shared/common/src/com/ngos/shared/common/elf/type.h
@@ -54,7 +54,7 @@ inline const char8* enumToFullString(ElfType type) // TEST: NO
 
     static char8 res[25];
 
-    sprintf(res, "0x%04X (%s)", type, enumToString(type));
+    sprintf(res, "0x%04X (%s)", (u16)type, enumToString(type));
 
     return res;
 }
