+++ src/os/shared/common/src/com/ngos/shared/common/fpu/xfeature.h
@@ -61,7 +61,7 @@ inline const char8* enumToFullString(XFeature feature) // TEST: NO
 
     static char8 res[45];
 
-    sprintf(res, "0x%02X (%s)", feature, enumToString(feature));
+    sprintf(res, "0x%02X (%s)", (u8)feature, enumToString(feature));
 
     return res;
 }
