+++ src/os/shared/common/src/com/ngos/shared/common/graphics/jpeg/lib/jpegmarkertype.h
@@ -110,7 +110,7 @@ inline const char8* enumToFullString(JpegMarkerType type) // TEST: NO
 
     static char8 res[33];
 
-    sprintf(res, "0x%02X (%s)", type, enumToString(type));
+    sprintf(res, "0x%02X (%s)", (u8)type, enumToString(type));
 
     return res;
 }
