+++ src/os/shared/common/src/com/ngos/shared/common/graphics/png/lib/pngchunktype.h
@@ -58,7 +58,7 @@ inline const char8* enumToFullString(PngChunkType type) // TEST: NO
 
     static char8 res[19];
 
-    sprintf(res, "0x%08X (%s)", type, enumToString(type));
+    sprintf(res, "0x%08X (%s)", (u32)type, enumToString(type));
 
     return res;
 }
